+++ Shell/Makefile
@@ -1,10 +1,9 @@
 OBJS = \
     Parser.o \
-    LineEditor.o \
     main.o
 
 PROGRAM = Shell
 
-LIB_DEPS = Core
+LIB_DEPS = Core LineEdit
 
 include ../Makefile.common
